Typo corrected
First version of the new edition of the converter.
This version was started on Friday, 24.05.2024 (get_files() and write_json) and continued on Saturday, 25.05.2025 (the rest in main.py).
On Friday, we also thought about creating a class instead of a collection of functions (see converter.py), but this idea was discarded because no significant added value could be found for the converter by using classes.

The current version is basically the same as the first converter, but at the moment some things are still hard-coded and are done without user input (e.g. input and output path or the command line arguments).
Comments also still need to be added.
All in all, however, it is more structured and clearer.

The converter can already extract all models in a file with several models and save/translate them in separate files/models.